+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- gRPC: accept keepalive parameters for gRPC outbound configuration.
 - observability: deprecation of logging levels `applicationError` and `serverError` in favor
   of `serverError` and `clientError`.